+++ .gitignore
@@ -1,1 +1,3 @@
 .stack-work/
+dist-newstyle/
+dist/
