+++ crdt/crdt.cabal
@@ -20,12 +20,12 @@ source-repository head
     type: git
     location: https://github.com/cblp/crdt.git
 
-flag Cm
-    description: Include commutative types
-    default: True
-
 library
     hs-source-dirs: lib
+    if impl(eta)
+        hs-source-dirs: lib-eta
+    else
+        hs-source-dirs: lib-ghc
     build-depends:  base >= 4.8 && < 4.11
                   , binary
                   , bytestring
@@ -33,11 +33,12 @@ library
                   , Diff
                   , hashable
                   , mtl
-                  , network-info
                   , safe
                   , stm
                   , time
                   , vector
+    if !impl(eta)
+        build-depends:  network-info
     exposed-modules:  CRDT.Cv
                       CRDT.Cv.GCounter
                       CRDT.Cv.GSet
@@ -52,14 +53,14 @@ library
                       CRDT.LWW
                       Data.MultiMap
                       Data.Semilattice
-    if flag(Cm)
-        cpp-options: -DENABLE_CM=1
+    if impl(ghc >= 8)
         exposed-modules:  CRDT.Cm
                           CRDT.Cm.Counter
                           CRDT.Cm.GSet
                           CRDT.Cm.ORSet
                           CRDT.Cm.RGA
                           CRDT.Cm.TwoPSet
+    other-modules: MacAddress
     default-language: Haskell2010
 
     if impl(ghc >= 8)
@@ -69,3 +70,10 @@ library
                       , semigroups
                       , transformers
         other-modules:  Compat
+
+executable crdt-getpid
+    default-language: Haskell2010
+    build-depends:  base 
+                  , crdt
+    hs-source-dirs: exe/crdt-getpid
+    main-is: Main.hs
