+++ crdt/exe/crdt-getpid/Main.hs
@@ -1,0 +1,4 @@
+import           CRDT.LamportClock (getMacAddress)
+
+main :: IO ()
+main = print =<< getMacAddress
