+++ crdt/lib/CRDT/LWW.hs
@@ -2,9 +2,9 @@
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE TypeFamilies #-}
 
-#if ENABLE_CM
+#if __GLASGOW_HASKELL__ >= 800
 {-# LANGUAGE LambdaCase #-}
-#endif /* ENABLE_CM */
+#endif /* __GLASGOW_HASKELL__ >= 800 */
 
 module CRDT.LWW
     ( LWW (..)
@@ -20,9 +20,9 @@ import           Data.Semigroup (Semigroup, (<>))
 
 import           Data.Semilattice (Semilattice)
 
-#if ENABLE_CM
+#if __GLASGOW_HASKELL__ >= 800
 import           CRDT.Cm (CausalOrd (..), CmRDT (..))
-#endif /* ENABLE_CM */
+#endif /* __GLASGOW_HASKELL__ >= 800 */
 
 import           CRDT.LamportClock (Clock, LamportTime (LamportTime), advance,
                                     getTime)
@@ -72,7 +72,7 @@ query = value
 --------------------------------------------------------------------------------
 -- CmRDT -----------------------------------------------------------------------
 
-#if ENABLE_CM
+#if __GLASGOW_HASKELL__ >= 800
 
 instance CausalOrd (LWW a) where
     precedes _ _ = False
@@ -91,7 +91,7 @@ instance Eq a => CmRDT (LWW a) where
         Just payload -> op <> payload
         Nothing      -> op
 
-#endif /* ENABLE_CM */
+#endif /* __GLASGOW_HASKELL__ >= 800 */
 
 advanceFromLWW :: Clock m => LWW a -> m ()
 advanceFromLWW LWW{time = LamportTime time _} = advance time
